Lvdeyang hoxton.sr9

查看合并请求 lvdeyang/tetris!733

+++ tetris-project/bvc/cloud/bvc2.0-web/src/main/java/com/sumavision/bvc/command/bandwidth/CommandStationBandwidthController.java
@@ -7,7 +7,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.sumavision.bvc.control.utils.UserUtils;
+import com.suma.venus.resource.service.BundleService;
+import com.sumavision.tetris.commons.exception.BaseException;
+import com.sumavision.tetris.commons.exception.code.StatusCode;
 import com.sumavision.tetris.commons.util.wrapper.HashMapWrapper;
 import com.sumavision.tetris.mvc.ext.response.json.aop.annotation.JsonBody;
 
@@ -16,13 +18,15 @@ import com.sumavision.tetris.mvc.ext.response.json.aop.annotation.JsonBody;
 public class CommandStationBandwidthController {
 
 	@Autowired
-	private UserUtils userUtils;
-	
-	@Autowired
 	private CommandStationBandwidthDAO commandStationBandwidthDao;
 	
 	@Autowired
 	private CommandStationBandwidthService commandStationBandwidthService;
+	
+	@Autowired
+	private BundleService bundleService;
+	
+	
 	/**
 	 * 添加站点<br/>
 	 * <b>作者:</b>lx<br/>
@@ -44,8 +48,9 @@ public class CommandStationBandwidthController {
 			Integer singleWidth,
 			String identity) throws Exception{
 		
-		CommandStationBandwidthPO station=new CommandStationBandwidthPO();
+		commandStationBandwidthService.alreadyExist(stationName, identity);
 		
+		CommandStationBandwidthPO station=new CommandStationBandwidthPO();
 		station.setStationName(stationName);
 		station.setSingleWidth(singleWidth);
 		station.setTotalWidth(totalWidth);
@@ -67,7 +72,14 @@ public class CommandStationBandwidthController {
 	@JsonBody
 	@ResponseBody
 	@RequestMapping(value="/delete")
-	public Object delete(Long id){
+	public Object delete(Long id)throws Exception{
+		
+		CommandStationBandwidthPO station = commandStationBandwidthDao.findOne(id);
+		
+		Boolean bundle = bundleService.checkedRegion(station.getIdentity());
+		if(bundle){
+			throw new BaseException(StatusCode.FORBIDDEN, "当前站点下存在设备，无法删除");
+		}
 		
 		commandStationBandwidthDao.delete(id);
 		
@@ -85,6 +97,7 @@ public class CommandStationBandwidthController {
 	 * @param singleWidth 单个带宽
 	 * @return
 	 * @throws ParseException
+	 * @throws BaseException 
 	 */
 	@JsonBody
 	@ResponseBody
@@ -94,14 +107,16 @@ public class CommandStationBandwidthController {
 			String stationName,
 			Integer totalWidth,
 			Integer singleWidth,
-			String identity) throws ParseException{
+			String identity) throws Exception{
+		
+		commandStationBandwidthService.alreadyExist(stationName, identity);
 
 		CommandStationBandwidthPO station=commandStationBandwidthDao.findOne(id);
 		
 		station.setStationName(stationName);
 		station.setSingleWidth(singleWidth);
 		station.setTotalWidth(totalWidth);
-		station.setIdentity(identity);
+//		station.setIdentity(identity);
 		
 		commandStationBandwidthDao.save(station);
 		
@@ -120,7 +135,7 @@ public class CommandStationBandwidthController {
 	@RequestMapping(value="/query")
 	public Object query(){
 		
-		commandStationBandwidthService.syncSerNodeToStation();
+//		commandStationBandwidthService.syncSerNodeToStation();
 		return new HashMapWrapper<String, Object>().put("rows", commandStationBandwidthDao.findAll()).getMap();
 		
 	}
